Move workspace route DTOs into owning modules
