+++ crates/server/src/bin/generate_types.rs
@@ -121,34 +121,34 @@ fn generate_types_content() -> String {
         server::routes::relay_auth::RefreshRelaySigningSessionResponse::decl(),
         server::routes::sessions::CreateFollowUpAttempt::decl(),
         server::routes::sessions::ResetProcessRequest::decl(),
-        server::routes::workspaces::ChangeTargetBranchRequest::decl(),
-        server::routes::workspaces::ChangeTargetBranchResponse::decl(),
-        server::routes::workspaces::AddWorkspaceRepoRequest::decl(),
-        server::routes::workspaces::AddWorkspaceRepoResponse::decl(),
-        server::routes::workspaces::MergeWorkspaceRequest::decl(),
-        server::routes::workspaces::PushWorkspaceRequest::decl(),
-        server::routes::workspaces::RenameBranchRequest::decl(),
-        server::routes::workspaces::RenameBranchResponse::decl(),
+        server::routes::workspaces::git::ChangeTargetBranchRequest::decl(),
+        server::routes::workspaces::git::ChangeTargetBranchResponse::decl(),
+        server::routes::workspaces::repos::AddWorkspaceRepoRequest::decl(),
+        server::routes::workspaces::repos::AddWorkspaceRepoResponse::decl(),
+        server::routes::workspaces::git::MergeWorkspaceRequest::decl(),
+        server::routes::workspaces::git::PushWorkspaceRequest::decl(),
+        server::routes::workspaces::git::RenameBranchRequest::decl(),
+        server::routes::workspaces::git::RenameBranchResponse::decl(),
         server::routes::sessions::review::StartReviewRequest::decl(),
         server::routes::sessions::review::ReviewError::decl(),
-        server::routes::workspaces::OpenEditorRequest::decl(),
-        server::routes::workspaces::OpenEditorResponse::decl(),
+        server::routes::workspaces::integration::OpenEditorRequest::decl(),
+        server::routes::workspaces::integration::OpenEditorResponse::decl(),
         db::models::requests::CreateWorkspaceApiRequest::decl(),
         db::models::requests::LinkedIssueInfo::decl(),
         server::routes::workspaces::pr::CreatePrApiRequest::decl(),
         server::routes::images::ImageResponse::decl(),
         server::routes::images::ImageMetadata::decl(),
         db::models::requests::WorkspaceRepoInput::decl(),
-        server::routes::workspaces::RunAgentSetupRequest::decl(),
-        server::routes::workspaces::RunAgentSetupResponse::decl(),
+        server::routes::workspaces::integration::RunAgentSetupRequest::decl(),
+        server::routes::workspaces::integration::RunAgentSetupResponse::decl(),
         server::routes::workspaces::gh_cli_setup::GhCliSetupError::decl(),
-        server::routes::workspaces::RebaseWorkspaceRequest::decl(),
-        server::routes::workspaces::ContinueRebaseRequest::decl(),
-        server::routes::workspaces::AbortConflictsRequest::decl(),
-        server::routes::workspaces::GitOperationError::decl(),
-        server::routes::workspaces::PushError::decl(),
+        server::routes::workspaces::git::RebaseWorkspaceRequest::decl(),
+        server::routes::workspaces::git::ContinueRebaseRequest::decl(),
+        server::routes::workspaces::git::AbortConflictsRequest::decl(),
+        server::routes::workspaces::git::GitOperationError::decl(),
+        server::routes::workspaces::git::PushError::decl(),
         server::routes::workspaces::pr::PrError::decl(),
-        server::routes::workspaces::RunScriptError::decl(),
+        server::routes::workspaces::execution::RunScriptError::decl(),
         server::routes::workspaces::images::AssociateWorkspaceImagesRequest::decl(),
         server::routes::workspaces::images::ImportIssueAttachmentsRequest::decl(),
         server::routes::workspaces::images::ImportIssueAttachmentsResponse::decl(),
@@ -167,7 +167,7 @@ fn generate_types_content() -> String {
         server::routes::workspaces::pr::CreateWorkspaceFromPrBody::decl(),
         server::routes::workspaces::pr::CreateWorkspaceFromPrResponse::decl(),
         server::routes::workspaces::pr::CreateFromPrError::decl(),
-        server::routes::workspaces::RepoBranchStatus::decl(),
+        server::routes::workspaces::git::RepoBranchStatus::decl(),
         db::models::requests::UpdateWorkspace::decl(),
         server::routes::workspaces::workspace_summary::WorkspaceSummaryRequest::decl(),
         server::routes::workspaces::workspace_summary::WorkspaceSummary::decl(),
