+++ crates/server/src/routes/workspaces/mod.rs
@@ -13,191 +13,13 @@ pub mod repos;
 pub mod streams;
 pub mod workspace_summary;
 
-use ::git::ConflictOp;
 use axum::{
     Router,
     middleware::from_fn_with_state,
     routing::{get, post},
 };
-use db::models::{merge::Merge, workspace::Workspace, workspace_repo::RepoWithTargetBranch};
-use serde::{Deserialize, Serialize};
-use ts_rs::TS;
-use uuid::Uuid;
 
 use crate::{DeploymentImpl, middleware::load_workspace_middleware};
-
-#[derive(Debug, Deserialize, Serialize, TS)]
-pub struct RebaseWorkspaceRequest {
-    pub repo_id: Uuid,
-    pub old_base_branch: Option<String>,
-    pub new_base_branch: Option<String>,
-}
-
-#[derive(Debug, Deserialize, Serialize, TS)]
-pub struct AbortConflictsRequest {
-    pub repo_id: Uuid,
-}
-
-#[derive(Debug, Deserialize, Serialize, TS)]
-pub struct ContinueRebaseRequest {
-    pub repo_id: Uuid,
-}
-
-#[derive(Debug, Serialize, Deserialize, TS)]
-#[serde(tag = "type", rename_all = "snake_case")]
-#[ts(tag = "type", rename_all = "snake_case")]
-pub enum GitOperationError {
-    MergeConflicts {
-        message: String,
-        op: ConflictOp,
-        conflicted_files: Vec<String>,
-        target_branch: String,
-    },
-    RebaseInProgress,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct DiffStreamQuery {
-    #[serde(default)]
-    pub stats_only: bool,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct WorkspaceStreamQuery {
-    pub archived: Option<bool>,
-    pub limit: Option<i64>,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct DeleteWorkspaceQuery {
-    #[serde(default)]
-    pub delete_remote: bool,
-    #[serde(default)]
-    pub delete_branches: bool,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct LinkWorkspaceRequest {
-    pub project_id: Uuid,
-    pub issue_id: Uuid,
-}
-
-#[derive(Debug, Deserialize, Serialize, TS)]
-pub struct AddWorkspaceRepoRequest {
-    pub repo_id: Uuid,
-    pub target_branch: String,
-}
-
-#[derive(Debug, Serialize, TS)]
-pub struct AddWorkspaceRepoResponse {
-    pub workspace: Workspace,
-    pub repo: RepoWithTargetBranch,
-}
-
-#[derive(Debug, Deserialize, Serialize, TS)]
-pub struct RunAgentSetupRequest {
-    pub executor_profile_id: executors::profile::ExecutorProfileId,
-}
-
-#[derive(Debug, Serialize, TS)]
-pub struct RunAgentSetupResponse {}
-
-#[derive(Debug, Deserialize, Serialize, TS)]
-pub struct MergeWorkspaceRequest {
-    pub repo_id: Uuid,
-}
-
-#[derive(Debug, Deserialize, Serialize, TS)]
-pub struct PushWorkspaceRequest {
-    pub repo_id: Uuid,
-}
-
-#[derive(Debug, Serialize, Deserialize, TS)]
-#[serde(tag = "type", rename_all = "snake_case")]
-#[ts(tag = "type", rename_all = "snake_case")]
-pub enum PushError {
-    ForcePushRequired,
-}
-
-#[derive(serde::Deserialize, TS)]
-pub struct OpenEditorRequest {
-    editor_type: Option<String>,
-    file_path: Option<String>,
-}
-
-#[derive(Debug, Serialize, TS)]
-pub struct OpenEditorResponse {
-    pub url: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, TS)]
-pub struct BranchStatus {
-    pub commits_behind: Option<usize>,
-    pub commits_ahead: Option<usize>,
-    pub has_uncommitted_changes: Option<bool>,
-    pub head_oid: Option<String>,
-    pub uncommitted_count: Option<usize>,
-    pub untracked_count: Option<usize>,
-    pub target_branch_name: String,
-    pub remote_commits_behind: Option<usize>,
-    pub remote_commits_ahead: Option<usize>,
-    pub merges: Vec<Merge>,
-    pub is_rebase_in_progress: bool,
-    pub conflict_op: Option<ConflictOp>,
-    pub conflicted_files: Vec<String>,
-    pub is_target_remote: bool,
-}
-
-#[derive(Debug, Clone, Serialize, TS)]
-pub struct RepoBranchStatus {
-    pub repo_id: Uuid,
-    pub repo_name: String,
-    #[serde(flatten)]
-    pub status: BranchStatus,
-}
-
-#[derive(serde::Deserialize, Debug, TS)]
-pub struct ChangeTargetBranchRequest {
-    pub repo_id: Uuid,
-    pub new_target_branch: String,
-}
-
-#[derive(serde::Serialize, Debug, TS)]
-pub struct ChangeTargetBranchResponse {
-    pub repo_id: Uuid,
-    pub new_target_branch: String,
-    pub status: (usize, usize),
-}
-
-#[derive(serde::Deserialize, Debug, TS)]
-pub struct RenameBranchRequest {
-    pub new_branch_name: String,
-}
-
-#[derive(serde::Serialize, Debug, TS)]
-pub struct RenameBranchResponse {
-    pub branch: String,
-}
-
-#[derive(Debug, Serialize, Deserialize, TS)]
-#[serde(tag = "type", rename_all = "snake_case")]
-#[ts(tag = "type", rename_all = "snake_case")]
-pub enum RenameBranchError {
-    EmptyBranchName,
-    InvalidBranchNameFormat,
-    OpenPullRequest,
-    BranchAlreadyExists { repo_name: String },
-    RebaseInProgress { repo_name: String },
-    RenameFailed { repo_name: String, message: String },
-}
-
-#[derive(Debug, Serialize, Deserialize, TS)]
-#[serde(tag = "type", rename_all = "snake_case")]
-#[ts(tag = "type", rename_all = "snake_case")]
-pub enum RunScriptError {
-    NoScriptConfigured,
-    ProcessAlreadyRunning,
-}
 
 pub fn router(deployment: &DeploymentImpl) -> Router<DeploymentImpl> {
     let workspace_id_router = Router::new()
